socrates_env/server/requirements.txt
openenv-core
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0
websockets>=11.0
pyyaml>=6.0
sentence-transformers>=2.2.0
numpy>=1.24.0
